requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
sentence-transformers
ollama
numpy
streamlit
requests
beautifulsoup4